Code Refactoring #2
