+++ src/components/HelperComponents/FormCreateModal/FormCreateModal.tsx
@@ -2,7 +2,7 @@ import {
   formElementStyling,
   handleFormSubTaskSubmit,
   handleFormSubmit,
-} from "../utils/formModalHelper";
+} from "../../../utils/formModalHelper";
 
 export default function FormCreateModal({
   title,
