+++ src/components/HelperComponents/SubTask/CreateSubTaskModal.tsx
@@ -1,24 +1,26 @@
 import React, { useState } from "react";
 import { useMutation } from "@apollo/client";
 
-import FormCreateModal from "../FormCreateModal";
-import CloseCreateModal from "../CloseCreateModal";
+import FormCreateModal from "../FormCreateModal/FormCreateModal";
+import CloseCreateModal from "../FormCreateModal/CloseCreateModal";
 
-import { CREATE_TASK_MUTATION, GET_ALL_TASKS } from "../../pages/api/crud_task";
+import { CREATE_SUBTASK_MUTATION } from "../../../pages/api/crud_subtask";
 
-export default function CreateTaskModal({ onClose }) {
+export default function CreateSubTaskModal({
+  onClose,
+  taskId,
+  allSubTasksRefetch,
+}) {
   const [title, setTitle] = useState("");
   const [description, setDescription] = useState("");
   const [status, setStatus] = useState("PENDING");
-  const [createTask] = useMutation(CREATE_TASK_MUTATION, {
-    refetchQueries: [{ query: GET_ALL_TASKS }],
-  });
+  const [createTask] = useMutation(CREATE_SUBTASK_MUTATION);
 
   return (
     <div className="fixed inset-0 bg-gray-600 bg-opacity-75 flex justify-center items-center">
       <div className="relative flex flex-col space-y-3 bg-white rounded-lg shadow p-8 w-full max-w-md">
         <CloseCreateModal onClose={onClose} />
-        <div className="text-lg font-bold">Create a task</div>
+        <div className="text-lg font-bold">Create a subtask</div>
         <FormCreateModal
           title={title}
           description={description}
@@ -28,6 +30,8 @@ export default function CreateTaskModal({ onClose }) {
           setStatus={setStatus}
           createTask={createTask}
           onClose={onClose}
+          taskId={taskId}
+          allSubTasksRefetch={allSubTasksRefetch}
         />
       </div>
     </div>
