+++ src/components/HelperComponents/SubTask/ShowCreateSubTaskModal.tsx
@@ -1,0 +1,17 @@
+import CreateSubTaskModal from "./CreateSubTaskModal";
+
+export default function ShowCreateSubTaskModal({
+  setIsModalOpen,
+  task,
+  allSubTasksRefetch,
+}) {
+  return (
+    <CreateSubTaskModal
+      onClose={() => {
+        setIsModalOpen(false);
+      }}
+      taskId={task.id}
+      allSubTasksRefetch={allSubTasksRefetch}
+    />
+  );
+}
