+++ src/components/HomePage/HeaderTaskInfo.tsx
@@ -1,4 +1,4 @@
-export default function HomeHeaderItem({ icon: Icon, name, value }) {
+export default function HeaderTaskInfo({ icon: Icon, name, value }) {
   return (
     <div className="relative flex flex-col w-52 h-28 rounded-lg bg-slate-500 text-white">
       <div className="flex items-center gap-2 pl-2 pt-2">
