+++ src/components/InfoPage/InfoHeader.tsx
@@ -1,0 +1,14 @@
+export default function InfoHeader() {
+  return (
+    <div className="relative flex flex-col space-y-10 p-6 w-full bg-slate-500 rounded-lg">
+      <div></div>
+      <div className="text-4xl">Info</div>
+      <div className="flex items-center space-x-6">
+        <div className=""></div>
+        <div>Created on</div>
+      </div>
+
+      <div className=""></div>
+    </div>
+  );
+}
