+++ src/components/InfoPage/InforBody.tsx
@@ -1,0 +1,3 @@
+export default function InfoBody() {
+  return <div className="relative bg-slate-500 p-6 rounded-lg"></div>;
+}
