+++ src/models/props/DeleteButtonProps.ts
@@ -1,0 +1,9 @@
+import { SubTask } from "../SubTaskModel";
+import { Task } from "../TaskModel";
+
+export interface DeleteButtonProps {
+  item: Task | SubTask;
+  item_type: string;
+  deleteItem;
+  allSubTasksRefetch?;
+}
