+++ src/pages/info/index.tsx
@@ -1,7 +1,3 @@
 export default function Info() {
-  return (
-    <div>
-      <h1>Info Page</h1>
-    </div>
-  );
+  return <div></div>;
 }
