+++ src/pages/signout/index.tsx
@@ -1,7 +1,9 @@
+import styles from "../../styles/Home.module.css";
+
 export default function SignOut() {
   return (
-    <div>
-      <h1>Sign Out</h1>
+    <div className="absolute top-80 right-32">
+      <h1 className={styles.title}>Only for design purposes for SideNav</h1>
     </div>
   );
 }
